refactor routing for each elasticms config
The default symfony router will no be override be a chain router.
This allows use to tag new routers for each elasticms config.

Now routers will be fully loaded based on the environment variables,
this was not possible before because symfony routes are always static!

In the feature the new router can also generate the ems urls!

Info:
- https://symfony.com/doc/current/cmf/book/routing.html
- https://symfony.com/doc/current/cmf/bundles/routing/dynamic.html