refactor: Routing transform clientRequestManager
The compiler should not inject the clientRequest, we can just use manager->getDefault(). In the feature the clientRequestManager will also be removed.

About the ems_client_helper.routing.relative_paths, it is not configurable from our current repositories. We don't use the option anymore. So we removed it, because it was the last option in the routing section.